refactor: onSuccess 로직을 사용부 컴포넌트에서 선언해주도록 수정
