+++ src/api/endpoint/feed/editFeed.ts
@@ -1,12 +1,10 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useRouter } from 'next/router';
-import { playgroundLink } from 'playground-common/export';
+import { useMutation } from '@tanstack/react-query';
+// import { playgroundLink } from 'playground-common/export';
 import { z } from 'zod';
 
-import { getPost } from '@/api/endpoint/feed/getPost';
-import { useGetPostsInfiniteQuery } from '@/api/endpoint/feed/getPosts';
+// import { getPost } from '@/api/endpoint/feed/getPost';
+// import { useGetPostsInfiniteQuery } from '@/api/endpoint/feed/getPosts';
 import { createEndpoint } from '@/api/typedAxios';
-import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 
 interface RequestBody {
   postId: number;
@@ -27,18 +25,8 @@ export const editFeed = createEndpoint({
   serverResponseScheme: z.unknown(),
 });
 
-export const useSaveEditFeedData = (feedId: string) => {
-  const router = useRouter();
-  const queryClient = useQueryClient();
-  const { logSubmitEvent } = useEventLogger();
-
+export const useSaveEditFeedData = () => {
   return useMutation({
     mutationFn: (reqeustBody: RequestBody) => editFeed.request(reqeustBody),
-    onSuccess: async () => {
-      logSubmitEvent('editCommunity');
-      queryClient.invalidateQueries({ queryKey: useGetPostsInfiniteQuery.getKey('') });
-      queryClient.invalidateQueries({ queryKey: getPost.cacheKey(feedId) });
-      await router.push(playgroundLink.feedList());
-    },
   });
 };
